% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputTestFunctions.R
\name{checkBetween0and1}
\alias{checkBetween0and1}
\title{checkBetween0and1}
\usage{
checkBetween0and1(x)
}
\description{
Stops execution with an error message if the input is not between 0 and 1
}
\keyword{internal}
