% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputTestFunctions.R
\name{checkDimensionsMatch}
\alias{checkDimensionsMatch}
\title{checkDimensionsMatch}
\usage{
checkDimensionsMatch(x, y)
}
\description{
Stops execution with an error message if the length of x is not equal to either 1 or the length of y
}
\keyword{internal}
