% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRPrepModel.R
\name{checkInputs.PrEP}
\alias{checkInputs.PrEP}
\title{Check PrEP inputs}
\usage{
checkInputs.PrEP(
  populationFractions,
  prepDiagnosisRate = 0,
  fractionOnPrEP = 0,
  timeBetweenPrEPDoses = 1,
  AVEs = 0,
  AVEi = 0,
  AVEp = 0,
  ...
)
}
\value{
List of PrEP parameters
}
\description{
Checks PrEP inputs and computes the parameters
}
\keyword{internal}
