% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRPrepModel.R
\name{checkInputs.SEIRPrEP}
\alias{checkInputs.SEIRPrEP}
\title{Check SEIR+V inputs}
\usage{
checkInputs.SEIRPrEP(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  prepDiagnosisRate,
  fractionOnPrEP,
  timeBetweenPrEPDoses,
  AVEs,
  AVEi,
  AVEp,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)
}
\value{
List of parameters for the SEIR+V model
}
\description{
Checks the input parameters for the SEIR+V model
}
\keyword{internal}
