% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRTModel.R
\name{checkInputs.SEIRT}
\alias{checkInputs.SEIRT}
\title{Check SEIR+T inputs}
\usage{
checkInputs.SEIRT(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  fractionSymptomatic,
  fractionSeekCare,
  fractionDiagnosedAndPrescribedOutpatient,
  fractionAdhere,
  fractionAdmitted,
  fractionDiagnosedAndPrescribedInpatient,
  AVEi,
  AVEp,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)
}
\value{
List of parameters for the SEIR+T model
}
\description{
Checks the input parameters for the SEIR+T model
}
\keyword{internal}
