% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRTVModel.R
\name{checkInputs.SEIRTV}
\alias{checkInputs.SEIRTV}
\title{Check SEIR+TV inputs}
\usage{
checkInputs.SEIRTV(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  fractionSymptomatic,
  fractionSeekCare,
  fractionDiagnosedAndPrescribedOutpatient,
  fractionAdhere,
  fractionAdmitted,
  fractionDiagnosedAndPrescribedInpatient,
  AVEi,
  AVEp,
  vaccineAdministrationRatePerDay,
  vaccineAvailabilityByDay,
  vaccineUptakeMultiplier,
  VEs,
  VEi,
  VEp,
  vaccineEfficacyDelay,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)
}
\value{
List of parameters for the SEIR+V model
}
\description{
Checks the input parameters for the SEIR+V model
}
\keyword{internal}
