% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRTV2DoseModel.R
\name{checkInputs.SEIRTV2Dose}
\alias{checkInputs.SEIRTV2Dose}
\title{Check SEIR+V 2-dose inputs}
\usage{
checkInputs.SEIRTV2Dose(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  fractionSymptomatic,
  fractionSeekCare,
  fractionDiagnosedAndPrescribedOutpatient,
  fractionAdhere,
  fractionAdmitted,
  fractionDiagnosedAndPrescribedInpatient,
  AVEi,
  AVEp,
  vaccineAdministrationRatePerDay,
  vaccineAvailabilityByDay,
  vaccineUptakeMultiplier,
  dose2Delay,
  VEs1,
  VEs2,
  VEi1,
  VEi2,
  VEp1,
  VEp2,
  vaccineEfficacyDelay,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)
}
\value{
List of parameters for the SEIR+V 2-dose model
}
\description{
Checks the input parameters for the SEIR+V 2-dose model
}
\keyword{internal}
