% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRVModel.R, R/SEIRVModel2.R
\name{checkInputs.SEIRV}
\alias{checkInputs.SEIRV}
\title{Check SEIR+V inputs}
\usage{
checkInputs.SEIRV(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  vaccineAdministrationRatePerDay,
  vaccineAvailabilityByDay,
  vaccineUptakeMultiplier,
  VEs,
  VEi,
  VEp,
  vaccineEfficacyDelay,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)

checkInputs.SEIRV(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  vaccineAdministrationRatePerDay,
  vaccineAvailabilityByDay,
  vaccineUptakeMultiplier,
  VEs,
  VEi,
  VEp,
  vaccineEfficacyDelay,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)
}
\value{
List of parameters for the SEIR+V model

List of parameters for the SEIR+V model
}
\description{
Checks the input parameters for the SEIR+V model

Checks the input parameters for the SEIR+V model
}
\keyword{internal}
