% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRVMonoModel.R
\name{checkInputs.SEIRVMono}
\alias{checkInputs.SEIRVMono}
\title{Check SEIR+VM inputs}
\usage{
checkInputs.SEIRVMono(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  vaccineAdministrationRatePerDay,
  vaccineAvailabilityDose1ByDay,
  vaccineAvailabilityDoseMByDay,
  vaccineUptakeMultiplier,
  timeOfExclusiveMonovalent,
  VEs1,
  VEsM,
  VEi1,
  VEiM,
  VEp1,
  VEpM,
  vaccineEfficacyDelay,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)
}
\value{
List of parameters for the SEIR+V model
}
\description{
Checks the input parameters for the SEIR+V model
}
\keyword{internal}
