% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRVPrepModel.R
\name{checkInputs.SEIRVPrEP}
\alias{checkInputs.SEIRVPrEP}
\title{Check SEIRVPrEP inputs}
\usage{
checkInputs.SEIRVPrEP(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  prepDiagnosisRate,
  fractionOnPrEP,
  timeBetweenPrEPDoses,
  AVEs,
  AVEi,
  AVEp,
  vaccineAdministrationRatePerDay,
  vaccineAvailabilityByDay,
  vaccineUptakeMultiplier,
  VEs,
  VEi,
  VEp,
  vaccineEfficacyDelay,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)
}
\value{
List of parameters for the SEIRVPrEP model
}
\description{
Checks the input parameters for the SEIRVPrEP model
}
\keyword{internal}
