% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRVModel.R, R/SEIRVModel2.R
\name{checkInputs.Vaccine}
\alias{checkInputs.Vaccine}
\title{Check vaccine inputs}
\usage{
checkInputs.Vaccine(
  population,
  populationFractions,
  seedStartDay,
  simulationLength,
  vaccineAdministrationRatePerDay = 0,
  vaccineAvailabilityByDay = 0,
  vaccineUptakeMultiplier = 1,
  VEs = 0,
  VEi = 0,
  VEp = 0,
  vaccineEfficacyDelay = 7,
  ...
)

checkInputs.Vaccine(
  population,
  populationFractions,
  seedStartDay,
  simulationLength,
  vaccineAdministrationRatePerDay = 0,
  vaccineAvailabilityByDay = 0,
  vaccineUptakeMultiplier = 1,
  VEs = 0,
  VEi = 0,
  VEp = 0,
  vaccineEfficacyDelay = 7,
  ...
)
}
\value{
List of vaccine parameters

List of vaccine parameters
}
\description{
Checks vaccine inputs and computes the vaccine parameters

Checks vaccine inputs and computes the vaccine parameters
}
\keyword{internal}
