% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRV2DoseModel.R
\name{checkInputs.Vaccine2Dose}
\alias{checkInputs.Vaccine2Dose}
\title{Check 2-Dose vaccine inputs}
\usage{
checkInputs.Vaccine2Dose(
  population,
  populationFractions,
  seedStartDay,
  simulationLength,
  vaccineAdministrationRatePerDay = 0,
  vaccineAvailabilityByDay = 0,
  vaccineUptakeMultiplier = 1,
  dose2Delay = 0,
  VEs1 = 0,
  VEs2 = 0,
  VEi1 = 0,
  VEi2 = 0,
  VEp1 = 0,
  VEp2 = 0,
  vaccineEfficacyDelay = 7,
  ...
)
}
\value{
List of 2-Dose vaccine parameters
}
\description{
Checks 2-Dose vaccine inputs and computes the vaccine parameters
}
\keyword{internal}
