% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputTestFunctions.R
\name{checkNonNegative}
\alias{checkNonNegative}
\title{checkNonNegative}
\usage{
checkNonNegative(x)
}
\description{
Stops execution with an error message if the input is not non-negative
}
\keyword{internal}
