% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputTestFunctions.R
\name{checkNonNegativeNumber}
\alias{checkNonNegativeNumber}
\title{checkNonNegativeNumber}
\usage{
checkNonNegativeNumber(x)
}
\description{
Stops execution with an error message if the input is not a non-negative number
}
\keyword{internal}
