% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputTestFunctions.R
\name{checkPositiveNumber}
\alias{checkPositiveNumber}
\title{checkPositiveNumber}
\usage{
checkPositiveNumber(x)
}
\description{
Stops execution with an error message if the input is not a positive number
}
\keyword{internal}
