% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActualRTimeSeries.R
\name{computeActualRTimeSeries}
\alias{computeActualRTimeSeries}
\title{Compute actual reproductive number time series}
\usage{
computeActualRTimeSeries(model, reconstructState, getDerivative)
}
\arguments{
\item{model}{A flumodels model object}

\item{reconstructState}{A function that reconstructs the state of the model}

\item{getDerivative}{A function that computes the derivative of the model}
}
\description{
This is a helper function that implements the calculations in
  the various 'getActualRTimeSeries' model class-specific functions
}
\keyword{internal}
