% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeVEsp.R
\name{computeVEp}
\alias{computeVEp}
\title{computeVEp}
\usage{
computeVEp(VEsp, VEsRatio, VEpRatio, VEs)
}
\arguments{
\item{VEsp}{Vaccine efficacy for combined protection of and prevention of symptomatic illness in vaccinated individuals.
A single fraction or vector of fractions by population group.}

\item{VEsRatio}{Ratio of VEs / VEp. VEsRatio = -1 is equivilant to VEsRatio = infinity (VEp = 0). One of VEsRatio, VEpRatio,
and VEs must be specified.}

\item{VEpRatio}{of VEp / VEs. VEpRatio = -1 is equivilant to VEpRatio = infinity (VEs = 0). One of VEsRatio, VEpRatio,
and VEs must be specified.}

\item{VEs}{Vaccine efficacy: protection for vaccinated susceptible
individuals; single fraction or vector of fractions by population group;
One of VEsRatio, VEpRatio, and VEs must be specified.}
}
\value{
A numeric or vector of numerics of same length as VEsp.
}
\description{
Compute VEp (Vaccine efficacy: prevention of symptomatic illness in
  infected indivduals) for a given VEsp (computed VEp + VEs). The presumed relationship is:
  (1 - VEsp) = (1 - VEp) * (1 - VEs).
}
\author{
Matthew Clay <clay.matt@gmail.com>
}
