% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeVEsp.R
\name{computeVEs}
\alias{computeVEs}
\title{computeVEs}
\usage{
computeVEs(VEsp, VEsRatio, VEpRatio, VEp)
}
\arguments{
\item{VEsp}{Vaccine efficacy for combined protection of and prevention of symptomatic illness in vaccinated individuals.
A single fraction or vector of fractions by population group.}

\item{VEsRatio}{Ratio of VEs / VEp. VEsRatio = -1 is equivilant to VEsRatio = infinity (VEp = 0). One of VEsRatio, VEpRatio,
and VEs must be specified.}

\item{VEpRatio}{Ratio of VEp / VEs. VEpRatio = -1 is equivilant to VEpRatio = infinity (VEs = 0). One of VEsRatio, VEpRatio,
and VEs must be specified.}

\item{VEp}{Vaccine efficacy: prevention of symptomatic illness in
infected indivduals; single fraction or vector of fractions by population group.
One of VEsRatio, VEpRatio, and VEs must be specified.}
}
\value{
A numeric or vector of numerics of same length as VEsp.
}
\description{
Compute VEs (Vaccine efficacy: protection for vaccinated susceptible
  individuals) for a given VEsp (computed VEp + VEs). The presumed relationship is:
  (1 - VEsp) = (1 - VEp) * (1 - VEs).
}
\author{
Matthew Clay <clay.matt@gmail.com>
}
