% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeVEsp.R
\name{computeVEsp}
\alias{computeVEsp}
\title{computeVEsp}
\usage{
computeVEsp(VEp = 0, VEs = 0)
}
\arguments{
\item{VEp}{Vaccine efficacy: prevention of symptomatic illness in
infected indivduals; single fraction or vector of fractions by population group;
defaults to 0. One of VEsRatio, VEpRatio, and VEs must be specified.}

\item{VEs}{Vaccine efficacy: protection for vaccinated susceptible
individuals; single fraction or vector of fractions by population group;
defaults to 0. One of VEsRatio, VEpRatio, and VEs must be specified.}
}
\value{
A numeric or vector of numerics of same length as VEp & VEs.
}
\description{
Compute VEsp (Vaccine efficacy for combined protection of and prevention of symptomatic illness in
vaccinated individuals) for a given VEp and VEs.
}
\author{
Matthew Clay <clay.matt@gmail.com>
}
