% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActualRTimeSeries.R
\name{getActualRTimeSeries}
\alias{getActualRTimeSeries}
\alias{getActualRTimeSeries.SEIRModel}
\alias{getActualRTimeSeries.SEIRTModel}
\alias{getActualRTimeSeries.SEIRVModel}
\alias{getActualRTimeSeries.SEIRTVModel}
\alias{getActualRTimeSeries.SEIRV2DoseModel}
\alias{getActualRTimeSeries.SEIRTV2DoseModel}
\alias{getActualRTimeSeries.SEIRVMonoModel}
\alias{getActualRTimeSeries.SEIRVPrimeBoostModel}
\title{Get actual reproductive number time series}
\usage{
getActualRTimeSeries(model)

\method{getActualRTimeSeries}{SEIRModel}(model)

\method{getActualRTimeSeries}{SEIRTModel}(model)

\method{getActualRTimeSeries}{SEIRVModel}(model)

\method{getActualRTimeSeries}{SEIRTVModel}(model)

\method{getActualRTimeSeries}{SEIRV2DoseModel}(model)

\method{getActualRTimeSeries}{SEIRTV2DoseModel}(model)

\method{getActualRTimeSeries}{SEIRVMonoModel}(model)

\method{getActualRTimeSeries}{SEIRVPrimeBoostModel}(model)
}
\arguments{
\item{model}{A flumodels model object}
}
\description{
Gets the time series of the actual observed effective reproductive 
  numbers for the model.
}
\keyword{internal}
