% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCombinedTimeSeries.R
\name{getCombinedTimeSeries}
\alias{getCombinedTimeSeries}
\title{Get combined time series}
\usage{
getCombinedTimeSeries(model, compartments)
}
\arguments{
\item{model}{The model from which to get the data values}

\item{compartments}{The vector of compartment names that contain the values to combine}
}
\value{
The combined time series
}
\description{
This is a utility function to sum the raw data time series that correspond to the specified compartments
}
\keyword{internal}
