% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCombinedValue.R
\name{getCombinedValue}
\alias{getCombinedValue}
\title{Get combined values}
\usage{
getCombinedValue(model, compartments, index)
}
\arguments{
\item{model}{The model from which to get the data values}

\item{compartments}{The vector of compartment names that contain the values to combine}

\item{index}{The index in the array from which data is desired}
}
\value{
The combined value
}
\description{
This is a utility function to sum the raw data values that correspond to the specified compartments at the given index
}
\keyword{internal}
