% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompartments.R
\name{getCompartments}
\alias{getCompartments}
\alias{getCompartments.SEIRModel}
\alias{getCompartments.SEIRVModel}
\alias{getCompartments.SEIRV2DoseModel}
\alias{getCompartments.SEIRVMonoModel}
\alias{getCompartments.SEIRVPrimeBoostModel}
\title{Get compartments}
\usage{
getCompartments(model, type)

\method{getCompartments}{SEIRModel}(model, type)

\method{getCompartments}{SEIRVModel}(model, type)

\method{getCompartments}{SEIRV2DoseModel}(model, type)

\method{getCompartments}{SEIRVMonoModel}(model, type)

\method{getCompartments}{SEIRVPrimeBoostModel}(model, type)
}
\arguments{
\item{model}{The model for which compartment names are desired}

\item{type}{The type of the compartment}
}
\value{
A vector of compartment names
}
\description{
This is a utility function to obtain the names of the compartments with the specified type
}
\keyword{internal}
