% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDayOfPeak.R
\name{getDayOfPeak}
\alias{getDayOfPeak}
\title{Get day of peak}
\usage{
getDayOfPeak(model, ...)
}
\arguments{
\item{model}{A flumodels model object}

\item{...}{Other parameters for the class-specific functions}
}
\description{
Gets the day of peak infections from the model
}
\details{
This is the generic function, for class-specific versions see:
\itemize{
	\item \code{\link{getDayOfPeak.SEIRModel}} for SEIR-type models
}
}
