% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDayOfPeak.R
\name{getDayOfPeak.SEIRModel}
\alias{getDayOfPeak.SEIRModel}
\title{Get day of peak}
\usage{
\method{getDayOfPeak}{SEIRModel}(model, incidence = FALSE, symptomatic = FALSE)
}
\arguments{
\item{model}{The model from which to get the data}

\item{incidence}{If true, returns infection incidence, otherwise returns infection prevalence; defaults to FALSE}

\item{symptomatic}{Whether or not to only report symptomatic infections; defaults to FALSE}
}
\value{
The day of peak infections
}
\description{
Gets the day of peak infections from the model
}
\keyword{internal}
