% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeaths.R
\name{getDeaths}
\alias{getDeaths}
\title{Get deaths}
\usage{
getDeaths(model, ...)
}
\arguments{
\item{model}{A flumodels model object}

\item{...}{Other parameters for the class-specific functions}
}
\description{
Gets the total deaths from the given model
}
\details{
This is the generic function, for class-specific versions see:
\itemize{
 \item \code{\link{getDeaths.SEIRModel}} for SEIR-type models
 \item \code{\link{getDeaths.SEIRTModel}} for SEIRT-type models
}
}
