% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeaths.R
\name{getDeaths.SEIRTModel}
\alias{getDeaths.SEIRTModel}
\title{Get deaths}
\usage{
\method{getDeaths}{SEIRTModel}(
  model,
  byGroup = TRUE,
  asRate = FALSE,
  caseFatalityRatio,
  timeSeries = FALSE
)
}
\arguments{
\item{model}{The model from which to get the data}

\item{byGroup}{Whether or not to return data by population group; defaults to TRUE}

\item{asRate}{Whether to return results as a rate (fraction of population) or else a number; defaults to FALSE}

\item{caseFatalityRatio}{The fraction of cases that result in fatalities; single fraction or vector of fractions by population group; must be specified}

\item{timeSeries}{Whether to return a time series or a final cumulative value}
}
\value{
A vector of deaths or death rates
}
\description{
Gets the total deaths from the given model
}
\keyword{internal}
