% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHospitalizations.R
\name{getHospitalizations}
\alias{getHospitalizations}
\title{Get hospitalizations}
\usage{
getHospitalizations(model, ...)
}
\arguments{
\item{model}{A flumodels model object}

\item{...}{Other parameters for the class-specific functions}
}
\description{
Gets the total hospitalizations from the given model
}
\details{
This is the generic function, for class-specific versions see:
\itemize{
 \item \code{\link{getHospitalizations.SEIRModel}} for SEIR-type models
 \item \code{\link{getHospitalizations.SEIRTModel}} for SEIRT-type models
}
}
