% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHospitalizations.R
\name{getHospitalizations.SEIRTModel}
\alias{getHospitalizations.SEIRTModel}
\title{Get hospitalizations}
\usage{
\method{getHospitalizations}{SEIRTModel}(
  model,
  byGroup = TRUE,
  asRate = FALSE,
  caseHospitalizationRatio,
  timeSeries = FALSE
)
}
\arguments{
\item{model}{The model from which to get the data}

\item{byGroup}{Whether or not to return data by population group; defaults to TRUE}

\item{asRate}{Whether to return results as a rate (fraction of population) or else a number; defaults to FALSE}

\item{caseHospitalizationRatio}{The fraction of cases that are hospitalized; single fraction or vector of fractions by population group; must be specified}

\item{timeSeries}{Whether to return a time series or a final cumulative value}
}
\value{
A vector of hospitalizations or hospitalization rates
}
\description{
Gets the total hospitalizations from the given model
}
\keyword{internal}
