% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfections.R
\name{getInfections}
\alias{getInfections}
\title{Get infections}
\usage{
getInfections(model, ...)
}
\arguments{
\item{model}{A flumodels model object}

\item{...}{Other parameters for the class-specific functions}
}
\description{
Gets the total infections from the given model
}
\details{
This is the generic function, for class-specific versions see:
\itemize{
 \item \code{\link{getInfections.SEIRModel}} for SEIR-type models
 \item \code{\link{getInfections.SEIRVModel}} for SEIRV-type models
 \item \code{\link{getInfections.SEIRV2DoseModel}} for SEIRV2Dose-type models
 \item \code{\link{getInfections.SEIRTModel}} for SEIRT-type models
 \item \code{\link{getInfections.SEIRTVModel}} for SEIRTV-type models
 \item \code{\link{getInfections.SEIRTV2DoseModel}} for SEIRTV2Dose-type models
}
}
