% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfections.R
\name{getInfections.SEIRTV2DoseModel}
\alias{getInfections.SEIRTV2DoseModel}
\title{Get infections}
\usage{
\method{getInfections}{SEIRTV2DoseModel}(model, byGroup = TRUE, asRate = FALSE, symptomatic = FALSE)
}
\arguments{
\item{model}{The model from which to get the data}

\item{byGroup}{Whether or not to return data by population group; defaults to TRUE}

\item{asRate}{Whether to return results as a rate (fraction of population) or else a number; defaults to FALSE}

\item{symptomatic}{Whether or not to only report symptomatic infections; defaults to FALSE}
}
\value{
A vector of infections
}
\description{
Gets the total infections from the given model
}
\keyword{internal}
