% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfections.R
\name{getInfections.SEIRVPrimeBoostModel}
\alias{getInfections.SEIRVPrimeBoostModel}
\title{Get infections}
\usage{
\method{getInfections}{SEIRVPrimeBoostModel}(
  model,
  byGroup = TRUE,
  asRate = FALSE,
  symptomatic = FALSE,
  fractionSymptomatic
)
}
\arguments{
\item{model}{The model from which to get the data}

\item{byGroup}{Whether or not to return data by population group; defaults to TRUE}

\item{asRate}{Whether to return results as a rate (fraction of population) or else a number; defaults to FALSE}

\item{symptomatic}{Whether or not to only report symptomatic infections; defaults to FALSE}

\item{fractionSymptomatic}{The fraction of cases that are symptomatic; must be specified if symptomatic = TRUE}
}
\value{
A vector of infections
}
\description{
Gets the total infections from the given model
}
\keyword{internal}
