% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfectionTimeSeries.R
\name{getInfectionTimeSeries}
\alias{getInfectionTimeSeries}
\title{Get infection time series}
\usage{
getInfectionTimeSeries(model, ...)
}
\arguments{
\item{model}{A flumodels model object}

\item{...}{Other parameters for the class-specific functions}
}
\description{
Gets the time series of infections from the given model
}
\details{
This is the generic function, for class-specific versions see:
\itemize{
 \item \code{\link{getInfectionTimeSeries.SEIRModel}} for SEIR-type models
 \item \code{\link{getInfectionTimeSeries.SEIRVModel}} for SEIRV-type models
 \item \code{\link{getInfectionTimeSeries.SEIRV2DoseModel}} for SEIRV2Dose-type models
 \item \code{\link{getInfectionTimeSeries.SEIRTModel}} for SEIRT-type models
 \item \code{\link{getInfectionTimeSeries.SEIRTVModel}} for SEIRTV-type models
 \item \code{\link{getInfectionTimeSeries.SEIRTV2DoseModel}} for SEIRTV2Dose-type models
}
}
