% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfectionTimeSeries.R
\name{getInfectionTimeSeries.SEIRVPrimeBoostModel}
\alias{getInfectionTimeSeries.SEIRVPrimeBoostModel}
\title{Get infection time series}
\usage{
\method{getInfectionTimeSeries}{SEIRVPrimeBoostModel}(
  model,
  byGroup = TRUE,
  asRate = FALSE,
  incidence = FALSE,
  symptomatic = FALSE,
  fractionSymptomatic,
  byWeek = FALSE
)
}
\arguments{
\item{model}{The model from which to get the data}

\item{byGroup}{Whether or not to return data by population group; defaults to TRUE}

\item{asRate}{Whether to return results as a rate (fraction of population) or else a number; defaults to FALSE}

\item{incidence}{If true, returns infection incidence, otherwise returns infection prevalence; defaults to FALSE}

\item{symptomatic}{Whether or not to only report symptomatic infections; defaults to FALSE}

\item{fractionSymptomatic}{The fraction of cases that are symptomatic; must be specified if symptomatic = TRUE}

\item{byWeek}{If true, returns the output on a weekly, not daily basis; defaults to FALSE}
}
\value{
A matrix that contains the infections by simulation day (or week, if byWeek is selected)
}
\description{
Gets the time series of infections from the given model
}
\keyword{internal}
