% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRTimeSeries.R
\name{getInfectiousnessVectorTimeSeries}
\alias{getInfectiousnessVectorTimeSeries}
\alias{getInfectiousnessVectorTimeSeries.SEIRModel}
\alias{getInfectiousnessVectorTimeSeries.SEIRTModel}
\alias{getInfectiousnessVectorTimeSeries.SEIRVModel}
\alias{getInfectiousnessVectorTimeSeries.SEIRTVModel}
\alias{getInfectiousnessVectorTimeSeries.SEIRV2DoseModel}
\alias{getInfectiousnessVectorTimeSeries.SEIRTV2DoseModel}
\alias{getInfectiousnessVectorTimeSeries.SEIRVMonoModel}
\alias{getInfectiousnessVectorTimeSeries.SEIRVPrimeBoostModel}
\title{Get infectiousness vector time series}
\usage{
getInfectiousnessVectorTimeSeries(model)

\method{getInfectiousnessVectorTimeSeries}{SEIRModel}(model)

\method{getInfectiousnessVectorTimeSeries}{SEIRTModel}(model)

\method{getInfectiousnessVectorTimeSeries}{SEIRVModel}(model)

\method{getInfectiousnessVectorTimeSeries}{SEIRTVModel}(model)

\method{getInfectiousnessVectorTimeSeries}{SEIRV2DoseModel}(model)

\method{getInfectiousnessVectorTimeSeries}{SEIRTV2DoseModel}(model)

\method{getInfectiousnessVectorTimeSeries}{SEIRVMonoModel}(model)

\method{getInfectiousnessVectorTimeSeries}{SEIRVPrimeBoostModel}(model)
}
\arguments{
\item{model}{A flumodels model object}
}
\description{
Gets the time series of effective fractions of the population
  by age group that are infectious.
}
\keyword{internal}
