% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabels.R
\name{getLabels}
\alias{getLabels}
\title{Get labels}
\usage{
getLabels(compartmentName, compartmentLength)
}
\arguments{
\item{compartmentName}{The base string name of the compartment}

\item{compartmentLength}{The length of the vector compartment}
}
\description{
This is a utility function to obtain the numbered compartment names
associated with the specified compartment string
}
\keyword{internal}
