% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRTimeSeries.R
\name{getRTimeSeries}
\alias{getRTimeSeries}
\title{Get effective reproductive number time series}
\usage{
getRTimeSeries(model, actual = FALSE)
}
\arguments{
\item{model}{A flumodels model object}

\item{actual}{A boolean value that determines whether to calculate the actual observed
effective R value (actual = TRUE), or the theoretical effective R value.}
}
\value{
The effective R time series
}
\description{
Gets the time series of effective reproductive numbers for the model.
}
