% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRTimeSeries.R
\name{getSusceptibilityVectorTimeSeries}
\alias{getSusceptibilityVectorTimeSeries}
\alias{getSusceptibilityVectorTimeSeries.SEIRModel}
\alias{getSusceptibilityVectorTimeSeries.SEIRVModel}
\alias{getSusceptibilityVectorTimeSeries.SEIRV2DoseModel}
\alias{getSusceptibilityVectorTimeSeries.SEIRVMonoModel}
\title{Get susceptibility vector time series}
\usage{
getSusceptibilityVectorTimeSeries(model)

\method{getSusceptibilityVectorTimeSeries}{SEIRModel}(model)

\method{getSusceptibilityVectorTimeSeries}{SEIRVModel}(model)

\method{getSusceptibilityVectorTimeSeries}{SEIRV2DoseModel}(model)

\method{getSusceptibilityVectorTimeSeries}{SEIRVMonoModel}(model)
}
\arguments{
\item{model}{A flumodels model object}
}
\description{
Gets the time series of effective fractions of the population
  by age group that are susceptible to infection.
}
\keyword{internal}
