% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateModel.R
\name{integrateModel}
\alias{integrateModel}
\title{Integrate model}
\usage{
integrateModel(
  initialState,
  parameters,
  derivativeFunction,
  seedFunction,
  rootFunction,
  eventFunction,
  method = "lsoda"
)
}
\arguments{
\item{initialState}{The initial state of the model's ODE system prior to seeding}

\item{parameters}{The model's parameters}

\item{derivativeFunction}{A function the computes the derivatives for the model's ODE system}

\item{seedFunction}{A function that modifies the state to seed infections}

\item{rootFunction}{An optional function(t, state, parameters) that triggers the
eventFunction when any result component is 0}

\item{eventFunction}{An optional function(t, state, parameters) that modifies state
when triggered by the rootFunction}

\item{method}{Which integration method to use. Defaults to lsoda}
}
\value{
A matrix of deSolve ode output for the model
}
\description{
Calculates and returns the raw output from the deSolve ODE integration
}
\keyword{internal}
