% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfectionTimeSeries.R
\name{makeOutputFromDailyInfectionsByGroup}
\alias{makeOutputFromDailyInfectionsByGroup}
\title{Get infection time series}
\usage{
makeOutputFromDailyInfectionsByGroup(
  model,
  dailyInfectionsByGroup,
  byGroup,
  asRate,
  byWeek
)
}
\arguments{
\item{model}{The model from which to get the data}

\item{dailyInfectionsByGroup}{Calculated daily infection prevalence (either as incidence or prevalence)}

\item{byGroup}{Whether or not to return data by population group; defaults to TRUE}

\item{asRate}{Whether to return results as a rate (fraction of population) or else a number; defaults to FALSE}

\item{byWeek}{If true, returns the output on a weekly, not daily basis; defaults to FALSE}
}
\value{
A matrix that contains the infections by simulation day (or week, if byWeek is selected)
}
\description{
Gets the time series of infections from the given model
}
\author{
Matt Clay clay.matt@gmail.com
}
\keyword{internal}
