% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot}
\usage{
plot(x, y, ...)
}
\arguments{
\item{x}{The model to plot}

\item{y}{Unused}

\item{...}{Other parameters for the class-specific functions}
}
\description{
Plot the model
}
\details{
This is the generic function, for class-specific versions see:
\itemize{
 \item \code{\link{plot.SEIRModel}} for SEIR-type models
 \item \code{\link{plot.SEIRTModel}} for SEIRT-type models
}
}
