% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SEIRModel}
\alias{plot.SEIRModel}
\title{Plot}
\usage{
\method{plot}{SEIRModel}(
  x,
  y,
  ...,
  asRate = FALSE,
  incidence = FALSE,
  symptomatic = FALSE,
  fractionSymptomatic,
  normalizeRate = FALSE,
  populationLabels
)
}
\arguments{
\item{x}{The model to plot}

\item{y}{Unused}

\item{asRate}{Whether to plot rates (fractions of population) or else
numbers; defaults to FALSE unless model population is 1}

\item{incidence}{If true, plots incidence, otherwise plots prevalence;
defaults to FALSE}

\item{symptomatic}{Whether or not to only report symptomatic infections; defaults to FALSE}

\item{fractionSymptomatic}{The fraction of cases that are symptomatic; must be specified if symptomatic = TRUE}

\item{normalizeRate}{Whether or not to normalize rates by population;
defaults to FALSE}

\item{populationLabels}{Labels for the population groups; optional}
}
\description{
Plot the model
}
\keyword{internal}
