% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SEIRTModel}
\alias{plot.SEIRTModel}
\title{Plot}
\usage{
\method{plot}{SEIRTModel}(
  x,
  ...,
  asRate = FALSE,
  incidence = FALSE,
  symptomatic = FALSE,
  normalizeRate = FALSE,
  populationLabels
)
}
\arguments{
\item{x}{The model to plot}

\item{asRate}{Whether to plot rates (fractions of population) or else
numbers; defaults to FALSE unless model population is 1}

\item{incidence}{If true, plots incidence, otherwise plots prevalence;
defaults to FALSE}

\item{symptomatic}{Whether or not to only plot symptomatic infections; defaults to FALSE}

\item{normalizeRate}{Whether or not to normalize rates by population;
defaults to FALSE}

\item{populationLabels}{Labels for the population groups; optional}

\item{y}{Unused}
}
\description{
Plot the model
}
\keyword{internal}
