% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRTModel.R
\name{SEIRTModel}
\alias{SEIRTModel}
\title{SEIR+T Model}
\usage{
SEIRTModel(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  fractionSymptomatic,
  fractionSeekCare,
  fractionDiagnosedAndPrescribedOutpatient,
  fractionAdhere,
  fractionAdmitted,
  fractionDiagnosedAndPrescribedInpatient,
  AVEi,
  AVEp,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)
}
\arguments{
\item{population}{Size of population; defaults to 1}

\item{populationFractions}{Vector of population fractions (all non-negative, 
sum to 1); defaults to 1, representing a single population group}

\item{contactMatrix}{A matrix whose (row i, column j) entry denotes the 
number of potentially infectious contacts a single individual from group j 
has with individuals from group i each day; defaults to proportional mixing}

\item{R0}{Average number of secondary cases from a single infected individual
in a completely susceptible population; must be specified}

\item{latentPeriod}{Latent period in days; must be specified}

\item{infectiousPeriod}{Infectious period in days; must be specified}

\item{seedInfections}{Fraction of the population to seed with infections; 
single fraction or vector of fractions by population group; defaults to 0}

\item{priorImmunity}{Fraction of the population with prior immunity; single 
fraction, or vector of fractions by population group; defaults to 0}

\item{useCommunityMitigation}{Whether or not to use community mitigation
implemented by modulation of the contact matrix; defaults to FALSE}

\item{communityMitigationStartDay}{If using community mitigation, day of the
simulation on which to start mitigation; must be specified if applicable}

\item{communityMitigationDuration}{If using community mitigation, duration of
time during which mitigation is in effect; must be specified if applicable}

\item{communityMitigationMultiplier}{If using community mitigation, the
non-negative matrix of multipliers that will be used to modulate the contact
matrix by elementwise multiplication; must be specified if applicable}

\item{fractionSymptomatic}{Fraction of the infections that are symptomatic 
cases; single fraction or vector of fractions by population group;  defaults
to 0.5}

\item{fractionSeekCare}{Fraction of the symptomatic cases that seek medical
care; single fraction or vector of fractions by population group; defaults
to 0.6}

\item{fractionDiagnosedAndPrescribedOutpatient}{Fraction of the outpatient
medical care seeking cases that are diagnosed and prescribed antiviral
drugs; single fraction or vector of fractions by population group; defaults
to 0.7}

\item{fractionAdhere}{Fraction of the cases that are prescribed antiviral
drugs that adhere to the regimen; single fraction or vector of fractions by
population group; defaults to 0.8}

\item{fractionAdmitted}{Fraction of the cases that require hospitalization
that are admitted; single fraction or vector of fractions by population
group; defaults to 1}

\item{fractionDiagnosedAndPrescribedInpatient}{Fraction of the hospitalized
cases that are diagnosed and prescribed antiviral drugs; single fraction or
vector of fractions by population group; defaults to 1}

\item{AVEi}{Antiviral efficacy: prevention of transmission from infected 
individuals taking antiviral drugs; single fraction or vector of fractions
by population group; defaults to 0}

\item{AVEp}{Antiviral efficacy: probability that antiviral treatment averts 
hospitalization and/or death; single fraction or vector of fractions by
population group; defaults to 0}

\item{simulationLength}{Number of days to simulate after seeding infections; 
defaults to 240}

\item{seedStartDay}{Day on which to seed initial infections; defaults to 0}

\item{tolerance}{Absolute tolerance for numerical integration; defaults to 
1e-8}

\item{method}{Which integration method to use. Defaults to lsoda}
}
\value{
a SEIRTModel object
}
\description{
A model for influenza that uses a SEIR framework with age 
  structure and that allows for antiviral treatment
}
