% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRVPrepModel.R
\name{SEIRVPrEPModel}
\alias{SEIRVPrEPModel}
\title{SEIRVPrep Model}
\usage{
SEIRVPrEPModel(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  prepDiagnosisRate,
  fractionOnPrEP,
  timeBetweenPrEPDoses,
  AVEs,
  AVEi,
  AVEp,
  vaccineAdministrationRatePerDay,
  vaccineAvailabilityByDay,
  vaccineUptakeMultiplier,
  VEs,
  VEi,
  VEp,
  vaccineEfficacyDelay,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)
}
\arguments{
\item{population}{Size of population; defaults to 1}

\item{populationFractions}{Vector of population fractions (all non-negative, 
sum to 1); defaults to 1, representing a single population group}

\item{contactMatrix}{A matrix whose (row i, column j) entry denotes the 
number of potentially infectious contacts a single individual from group j 
has with individuals from group i each day; defaults to proportional mixing}

\item{R0}{Average number of secondary cases from a single infected individual
in a completely susceptible population; must be specified}

\item{latentPeriod}{Latent period in days; must be specified}

\item{infectiousPeriod}{Infectious period in days; must be specified}

\item{seedInfections}{Number of infections to seed; single number or vector
of numbers by population group; defaults to 0}

\item{priorImmunity}{Fraction of the population with prior immunity; single 
fraction, or vector of fractions by population group; defaults to 0}

\item{useCommunityMitigation}{Whether or not to use community mitigation
implemented by modulation of the contact matrix; defaults to FALSE}

\item{communityMitigationStartDay}{If using community mitigation, days of the
simulation on which each cycle of mitigation starts; must be specified 
sequentially in time if applicable}

\item{communityMitigationDuration}{If using community mitigation, duration of
time during which each cycle of mitigation is in effect (different cycles of
mitigation cannot overlap); must be specified if applicable}

\item{communityMitigationMultiplier}{The communityMitigationMultiplier is the set
 of multipliers that will be used to modulate the contact
  matrix by elementwise multiplication and can be:
(i) a single non-negative scalar, if mitigation is the same for all subgroups 
and all mitigation cycles; or, (ii) an atomic vector of non-negative entries 
with length equal to the number of mitigation rounds, if mitigation is the 
same for all subgroups but varies by  mitigation cycles; or
(iii) a single 2D array of non-negative entries whose dimensions are identical
 to those of the contact matrix, if mitigation varies by subgroup but is the 
 same across mitigation cycles; or (iv) a list of 2D arrays, each of which has
 dimensions that are identical to those of the contact matrix;
the length of the list must match the number of mitigation cycles. This is 
the most general case, in which mitigation varies by subgroup and by mitigation cycle.
Must be specified if applicable.}

\item{prepDiagnosisRate}{Rate at which PrEP is diagnosed; defaults to 0}

\item{fractionOnPrEP}{Fraction of the population that takes PrEP; single fraction 
or vector of fractions by population group;  defaults to 0.0}

\item{timeBetweenPrEPDoses}{Time (in days) elapsed between PrEP doses; single number or 
or vector by population group; defaults to 1.0}

\item{AVEs}{Antiviral efficacy: prevention of transmission to susceptible
individuals taking PrEP; single fraction or vector of fractions
by population group; defaults to 0}

\item{AVEi}{Antiviral efficacy: prevention of transmission from infected 
individuals taking PrEP; single fraction or vector of fractions
by population group; defaults to 0}

\item{AVEp}{Antiviral efficacy: probability that antiviral treatment averts 
hospitalization and/or death; single fraction or vector of fractions by
population group; defaults to 0
#' @param vaccineAdministrationRatePerDay Vaccine administration rate each day; 
defaults to 0}

\item{vaccineAvailabilityByDay}{Vector that contains the amount of vaccine 
available each day; defaults to 0}

\item{vaccineUptakeMultiplier}{Vector of multipliers that determines the
relative rate at which vaccine is given to each age group; defaults to
vaccine being allotted proportionally by population}

\item{VEs}{Vaccine efficacy: protection for vaccinated susceptible 
individuals; single fraction or vector of fractions by population group;
defaults to 0}

\item{VEi}{Vaccine efficacy: prevention of transmission from vaccinated 
infected individuals; single fraction or vector of fractions by population
group; defaults to 0}

\item{VEp}{Vaccine efficacy: prevention of symptomatic illness in
infected indivduals; single fraction or vector of fractions by population
group; defaults to 0}

\item{vaccineEfficacyDelay}{Delay in days between administration of dose and 
onset of protection; defaults to 7}

\item{simulationLength}{Number of days to simulate after seeding infections; 
defaults to 240}

\item{seedStartDay}{Day on which to seed initial infections; defaults to 0}

\item{tolerance}{Absolute tolerance for numerical integration; defaults to 
1e-8}

\item{method}{Which integration method to use. Defaults to lsoda}
}
\value{
a SEIRModel object
}
\description{
A model for influenza that uses a SEIR framework with age 
  structure that incorporates PrEP antivirals and vaccination
}
