% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRVPrimeBoostModel.R
\name{SEIRVPrimeBoostModel}
\alias{SEIRVPrimeBoostModel}
\title{SEIR+V Prime-Boost Model}
\usage{
SEIRVPrimeBoostModel(
  population,
  populationFractions,
  contactMatrix,
  R0,
  latentPeriod,
  infectiousPeriod,
  seedInfections,
  priorImmunity,
  useCommunityMitigation,
  communityMitigationStartDay,
  communityMitigationDuration,
  communityMitigationMultiplier,
  vaccineAdministrationRatePerDay,
  vaccineAvailabilityByDayPrime,
  vaccineAvailabilityByDayBoost,
  vaccineUptakeMultiplier,
  boostDelay,
  VEs1,
  VEs2,
  VEi1,
  VEi2,
  VEp1,
  VEp2,
  vaccineEfficacyDelay,
  simulationLength,
  seedStartDay,
  tolerance,
  method
)
}
\arguments{
\item{population}{Size of population; defaults to 1}

\item{populationFractions}{Vector of population fractions (all non-negative, 
sum to 1); defaults to 1, representing a single population group}

\item{contactMatrix}{A matrix whose (row i, column j) entry denotes the 
number of potentially infectious contacts a single individual from group j 
has with individuals from group i each day; defaults to proportional mixing}

\item{R0}{Average number of secondary cases from a single infected individual
in a completely susceptible population; must be specified}

\item{latentPeriod}{Latent period in days; must be specified}

\item{infectiousPeriod}{Infectious period in days; must be specified}

\item{seedInfections}{Fraction of the population to seed with infections; 
single fraction, or vector of fractions by population group; defaults to 0}

\item{priorImmunity}{Fraction of the population with prior immunity; single 
fraction, or vector of fractions by population group; defaults to 0}

\item{useCommunityMitigation}{Whether or not to use community mitigation
implemented by modulation of the contact matrix; defaults to FALSE}

\item{communityMitigationStartDay}{If using community mitigation, day of the
simulation on which to start mitigation; must be specified if applicable}

\item{communityMitigationDuration}{If using community mitigation, duration of
time during which mitigation is in effect; must be specified if applicable}

\item{communityMitigationMultiplier}{If using community mitigation, the
non-negative matrix of multipliers that will be used to modulate the contact
matrix by elementwise multiplication; must be specified if applicable}

\item{vaccineAdministrationRatePerDay}{Vaccine administration rate each day; 
defaults to 0}

\item{vaccineAvailabilityByDayPrime}{Vector that contains the amount of Prime vaccine 
available each day; defaults to 0}

\item{vaccineAvailabilityByDayBoost}{Vector that contains the amount of boost vaccine 
available each day; defaults to 0}

\item{vaccineUptakeMultiplier}{Vector of multipliers that determines the
relative rate at which vaccine is given to each age group; defaults to
vaccine being allotted proportionally by population}

\item{boostDelay}{Minimum days between prime and boost doses; defaults to 14}

\item{VEs1}{Vaccine efficacy: protection for vaccinated susceptible 
individuals after priming dose; single fraction or vector of fractions by
population group; defaults to 0}

\item{VEs2}{Vaccine efficacy: protection for vaccinated susceptible 
individuals after priming and boost doses; single fraction or vector of fractions by
population group; defaults to 0}

\item{VEi1}{Vaccine efficacy: prevention of transmission from vaccinated 
infected individuals after priming dose; single fraction or vector of fractions
by population group; defaults to 0}

\item{VEi2}{Vaccine efficacy: prevention of transmission from vaccinated 
infected individuals after priming and boost doses; single fraction or vector of fractions
by population group; defaults to 0}

\item{VEp1}{Vaccine efficacy: prevention of symptomatic illness in
infected indivduals after priming dose; single fraction or vector of fractions 
by population group; defaults to 0}

\item{VEp2}{Vaccine efficacy: prevention of symptomatic illness in
infected indivduals after priming and boost doses; single fraction or vector of fractions 
by population group; defaults to 0}

\item{vaccineEfficacyDelay}{Delay in days between administration of dose and 
onset of protection; defaults to 7}

\item{simulationLength}{Number of days to simulate after seeding infections; 
defaults to 240}

\item{seedStartDay}{Day on which to seed initial infections; defaults to 0}

\item{tolerance}{Absolute tolerance for numerical integration; defaults to 
1e-8}

\item{method}{Which integration method to use. Defaults to lsoda}
}
\value{
a SEIRVPrimeBoostModel object
}
\description{
A model for influenza that uses a SEIR framework with age 
  structure and that allows for vaccination with a prime-boost vaccine
}
